building user app as a product
