+++ user_app/auth.py
@@ -1,0 +1,54 @@
+import jwt
+from .lib.db_helper import *  
+from flask import Blueprint, g, redirect, render_template, request, session, url_for, current_app, jsonify
+
+
+auth = Blueprint("auth", __name__, url_prefix="/app")
+
+
+@auth.route('/', methods=['GET', 'POST'])
+def index():
+    return render_template('index.html')
+
+
+@auth.route('/register', methods=['POST',])
+def register():
+    username = request.form['username']
+    password = request.form['password']
+    email = request.form['email']
+    current_app.logger.info("Registering user")
+    register_user(username, password, email)
+    return serve_response(data="User registered successfully" , status_code=200)
+    
+
+@auth.route('/login', methods=['POST',])
+def login():
+    username = request.form['username']
+    password = request.form['password']
+    user = login_user(username, password)
+    current_app.logger.info("Logging in user")
+    return jsonify(user)
+    
+
+@auth.route('/logout', methods=['DELETE',])
+def logout():
+    session.clear() 
+    return redirect(url_for('index'))
+
+
+@auth.route('/delete', methods=['DELETE',])
+def delete():
+    username = request.form['username']
+    delete_user(username)     
+    return serve_response(data="User deleted", status_code=200)
+    
+
+@auth.before_app_request
+def load_logged_in_user():
+    user_id = session.get('user_id')
+    if user_id is None:
+        g.user = None
+    else:
+        g.user = get_db_x().execute(
+            'SELECT * FROM user WHERE id = ?', (user_id,)
+        ).fetchone()
